feature/global_det_mpmd (#610)
* Initial commit of changes for atmos stats grid2grid MPMD job output

* Updates for copying to COMOUT in MPMD jobs

* Update TMP_DIR

* Remove SENDCOM checks

* Add MPMD working directory changes for global_det atmos g2g stats assemble_data

* Switch order of files to check as input

* Fix variable name

* Add MPMD working directory changes for global_det atmos g2g stats generate_stats

* Add MPMD working directory changes for global_det atmos g2g stats gather_stats

* Move os import

* Standardized INPUT_BASE

* Add file check

* Rework MPMD directory writing for global_det atmos grid2grid stats

* Add MPMD directory work for atmos grid2obs stats reformat_data

* Add MPMD directory work for atmos grid2obs stats assemble_data

* Set TMP_DIR to MET_TMP_DIR

* Remove extra import glob

* Rename directory

* Fix file name

* Only print missing data for False found_input

* Fix file path

* Remove logging messages if restart

* Add MPMD directory work for atmos grid2obs stats generate_stats

* Add MPMD directory work for atmos grid2obs stats garther_stats

* Add err_chk

* Add \n

* Make enhancements for restart

* Remove making of MET_TMP_DIR for atmos stats

* Add MPMD directory work for atmos wmo stats reformat_data

* Update script purpose

* Fix restart

* Add MPMD directory work for atmos wmo stats assemble_data

* Add MPMD directory work for atmos wmo stats generate_stats

* Add MPMD directory work for atmos wmo stats gather_stats

* Add MPMD directory work for wave stats

* Add MPMD directory work for wave plots

* Add inital MPMD directory work for atmos grid2grid plots condense_stats

* Add additional MPMD directory work for atmos grid2grid plots condense_stats

* Add MPMD directory work for atmos grid2grid plots filter_stats

* Add inital MPMD directory work for atmos grid2grid plots make_plots

* Add MPMD directory work for atmos grid2grid plots tar_images

* Improve restart for condense_stats, filter_stats, and make_plots

* Improve restart for tar_images

* Remove wind shear script copy to DATA

* Add inital MPMD directory work for atmos grid2grid precip plots

* Final changes for atmos grid2grid precip plots

* Changes for changes for atmos grid2grid snow plots

* Changes for atmos grid2grid snow plots

* Remove global_det_atmos_plots_production_tof240.py

* Fix typo

* Changes for atmos grid2obs plots

* Changes for atmos wmo monthly

* Fixes for NOHRSC spatial map

* Add rstprod permissions for files in working dir

+++ ush/global_det/global_det_atmos_stats_wmo_reformat_cnvstat.py
@@ -22,20 +22,11 @@ import global_det_atmos_util as gda_util
 print("BEGIN: "+os.path.basename(__file__))
 
 # Read in environment variables
-COMIN = os.environ['COMIN']
-COMINgfs = os.environ['COMINgfs']
-SENDCOM = os.environ['SENDCOM']
-COMOUT = os.environ['COMOUT']
 DATA = os.environ['DATA']
-NET = os.environ['NET']
-RUN = os.environ['RUN']
-VERIF_CASE = os.environ['VERIF_CASE']
-STEP = os.environ['STEP']
-COMPONENT = os.environ['COMPONENT']
 valid_date = os.environ['valid_date']
-JOB_GROUP = os.environ['JOB_GROUP']
 obs_file = os.environ['obs_file']
 input_ascii2nc_file = os.environ['input_ascii2nc_file']
+job_num_work_dir = os.environ['job_num_work_dir']
 
 valid_date_dt = datetime.datetime.strptime(valid_date, '%Y%m%d%H')
 
@@ -64,7 +55,7 @@ ascii2nc_df_dict = {
     'Observation_Value': []
 }
 
-os.chdir(os.path.join(DATA, 'gdas_cnvstat'))
+os.chdir(job_num_work_dir)
 print(f"Working in {os.getcwd()}")
 
 for diag_var in list(diag_var_dict.keys()):
@@ -75,10 +66,19 @@ for diag_var in list(diag_var_dict.keys()):
         with tarfile.open(obs_file, 'r') as tf:
             print(f"Extracting {diag_var_zipfile} from {obs_file}")
             tf.extract(member=diag_var_zipfile)
-        with gzip.open(diag_var_zipfile, 'rb') as dvzf:
-            with open(diag_var_file, 'wb') as dvf:
-                print(f"Unzipping {diag_var_zipfile} to {diag_var_file}")
-                shutil.copyfileobj(dvzf, dvf)
+            if os.path.exists(diag_var_zipfile):
+                gda_util.run_shell_command(['chmod', '750', diag_var_zipfile])
+                gda_util.run_shell_command(['chgrp', 'rstprod',
+                                            diag_var_zipfile])
+                with gzip.open(diag_var_zipfile, 'rb') as dvzf:
+                    with open(diag_var_file, 'wb') as dvf:
+                        print(f"Unzipping {diag_var_zipfile} to {diag_var_file}")
+                        shutil.copyfileobj(dvzf, dvf)
+                        if os.path.exists(diag_var_file):
+                            gda_util.run_shell_command(['chmod', '750',
+                                                        diag_var_file])
+                            gda_util.run_shell_command(['chgrp', 'rstprod',
+                                                        diag_var_file])
     if gda_util.check_file_exists_size(diag_var_file):
         print(f"Processing {diag_var_dict[diag_var]} from {diag_var_file}")
         diag_var_nc = netcdf.Dataset(diag_var_file, 'r')
@@ -277,6 +277,8 @@ ascii2nc_df = ascii2nc_df[
 ]
 
 # Write out dataframe
+input_ascii2nc_file_dir = input_ascii2nc_file.rpartition('/')[0]
+gda_util.make_dir(input_ascii2nc_file_dir)
 print(f"Writing file to {input_ascii2nc_file}")
 ascii2nc_df.to_csv(
     input_ascii2nc_file, header=None, index=None, sep=' ', mode='w'
